third_party/blink/web_tests/ppapi/plugins/window-open-expected.txt

CONSOLE MESSAGE: Blink Deprecated Test Plugin: PLUGIN: WINDOW OPEN SUCCESS
This tests that window.open invoked by a plugin without a javascript context succeeds.
